pyproject.toml
[tool.poetry]
name = "quiggler"
version = "1.0.0"
description = "quiggler"
authors = ["nalch"]
license = "Apache-2.0"
readme = "README.rst"
classifiers = []

[tool.poetry.dependencies]
python = "^3.6"
networkx = "^2.4"
Django = "^3.0.5"
djangorestframework = "^3.11.0"
markdown = "^3.2.1"
django-filter = "^2.2.0"
coreapi = "^2.3.3"
Pillow = "^7.1.2"
boltons = "^20.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.3."
ipython = "^7.11.1"
pytest-xdist = "^1.31.0"
pytest-cov = "^2.8.1"
pytest-mock = "^2.0.0"
pytest-pythonpath = "^0.7.3"
pydocstyle = "^5.0.2"
flake8 = "^3.7.9"
isort = "^4.3.21"
unify = "^0.5"
mypy = "^0.761"
pytype = "^2020.1.24"
Sphinx = "^2.3.1"
sphinx-rtd-theme = "^0.4.3"
black = "^19.10b0"
bump2version = "^1.0.0"
gitchangelog = "^3.0.4"

[tool.poetry.scripts]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
